% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobi.R
\name{cobi}
\alias{cobi}
\title{Print a sentence}
\usage{
cobi(sentence)
}
\arguments{
\item{sentence}{A character vector with one element.}
}
\value{
A character vector.
}
\description{
Print a sentence
}
\examples{
sentence <- "My sentence."
cobi(sentence)
}
