% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cobi_approp_line_categories}
\alias{cobi_approp_line_categories}
\title{COBI 'Line_Item_Cat_Desc' Dictionary}
\format{
\subsection{\code{cobi_approp_line_categories}}{

A data frame with 9 rows and 3 columns:
\describe{
\item{Line_Item_Cat_Desc}{value of 'Line_Item_Cat_Desc' field in COBI appropriations extract}
\item{COBI Org Alias}{alias ID used by LFA for joining with other tables}
\item{Definition}{definition of field value}
}
}
}
\source{
\url{https://cobi.utah.gov/2025/1/financials}
}
\usage{
cobi_approp_line_categories
}
\description{
Reference table documenting the 'Line_Item_Cat_Desc' field in COBI appropriations extracts returned by corresponding get_cobi_* functions, as defined on COBI
}
\keyword{datasets}
