% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobi_approps.R
\name{cobi_approps}
\alias{cobi_approps}
\title{Pull COBI appropriations}
\usage{
cobi_approps(years)
}
\arguments{
\item{years}{A vector of integers indicated year(s) of interest}
}
\value{
A tibble (modern data frame)
}
\description{
Pull COBI appropriations
}
\examples{
cobi_data <- cobi_approps(c(2024, 2025))
}
