% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cobi_approps_dictionary}
\alias{cobi_approps_dictionary}
\title{COBI Appropriations Data Dictionary}
\format{
\subsection{\code{cobi_approps_dictionary}}{

A data frame with 27 rows and 2 columns:
\describe{
\item{Field}{field name in COBI appropriations extract}
\item{Definition}{definition of field name}
}
}
}
\source{
Developed by Jackson Bolos (GOPB) and Brian Fay (LFA)
}
\usage{
cobi_approps_dictionary
}
\description{
Reference table documenting COBI appropriations extracts returned by *_cobi_approps functions
}
\keyword{datasets}
