% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi.R
\name{get_cobi_approps}
\alias{get_cobi_approps}
\title{Pull COBI appropriations}
\usage{
get_cobi_approps(
  years,
  agencies = NULL,
  line_items = NULL,
  appr_units = NULL,
  financing_sources = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{appr_units}{character vector indicating appropriation unit(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}
}
\value{
tibble (modern data frame)
}
\description{
Pull COBI appropriations
}
\examples{
cobi_data <- get_cobi_approps(2025)
cobi_data <- get_cobi_approps(years = 2025, appr_units = 'KPBAC', financing_sources = '9750')
cobi_data <- get_cobi_approps(years = c(2024, 2025), line_items = 'CBAA')
cobi_data <- get_cobi_approps(years = 2025, line_items = '2105', financing_sources = '9800')
cobi_data <- get_cobi_approps(years = c(2024, 2025), agencies = '060', financing_sources = '1000')
}
