% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi_data.R
\name{get_cobi_bla}
\alias{get_cobi_bla}
\title{Get COBI Business-like Activities}
\usage{
get_cobi_bla(
  years,
  agencies = NULL,
  line_items = NULL,
  financing_sources = NULL,
  is1x = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
tibble (modern data frame)
}
\description{
Get COBI Business-like Activities
}
\examples{
goeo_federal_bla <- get_cobi_bla(2025, agencies = '063', financing_sources = '9750')
}
