% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi_ref.R
\name{get_cobi_expenditure_categories}
\alias{get_cobi_expenditure_categories}
\title{Pull COBI expenditure categories}
\usage{
get_cobi_expenditure_categories()
}
\value{
tibble (modern data frame)
}
\description{
Pull COBI expenditure categories
}
\examples{
expenditure_categories <- get_cobi_expenditure_categories()
}
