% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi.R
\name{get_cobi_financing_sources}
\alias{get_cobi_financing_sources}
\title{Pull COBI financing sources}
\usage{
get_cobi_financing_sources(year = NULL, restricted_only = NULL)
}
\arguments{
\item{year}{integer indicating fiscal year time stamp at which to snapshot active financing sources}

\item{restricted_only}{boolean value indicating whether to pull only restricted funds/accounts}
}
\value{
tibble (modern data frame)
}
\description{
Pull COBI financing sources
}
\examples{
financing_sources <- get_cobi_financing_sources()
restricted_2024 <- get_cobi_financing_sources(2024, TRUE)
}
