% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi_data.R
\name{get_cobi_meta}
\alias{get_cobi_meta}
\title{Pull COBI meta}
\usage{
get_cobi_meta(
  years,
  agencies = NULL,
  line_items = NULL,
  appr_units = NULL,
  categories = NULL,
  is1x = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{appr_units}{character vector indicating appropriation unit(s) of interest}

\item{categories}{character vector indicating financing source(s) or expenditure category(ies) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
tibble (modern data frame)
}
\description{
Pull COBI meta
}
\examples{
cobi_meta <- get_cobi_meta(2025)
cobi_meta <- get_cobi_meta(years = 2025, appr_units = 'KPBAC', categories = '9750')
cobi_meta <- get_cobi_meta(years = c(2024, 2025), line_items = 'CBAA')
cobi_meta <- get_cobi_meta(years = 2025, line_items = '2105', categories = '9800')
cobi_meta <- get_cobi_meta(years = 2025, agencies = '060', categories = '1000')
cobi_meta <- get_cobi_meta(years = 2025, categories = 'AA')
}
