% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi.R
\name{get_cobi_orgs}
\alias{get_cobi_orgs}
\title{Pull COBI orgs}
\usage{
get_cobi_orgs(year = NULL, agencies = NULL, line_items = NULL)
}
\arguments{
\item{year}{integer indicating fiscal year time stamp at which to snapshot the budget organization hierarchy}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}
}
\value{
tibble (modern data frame)
}
\description{
Pull COBI orgs
}
\examples{
orgs <- get_cobi_orgs()
dhhs_orgs <- get_cobi_orgs(agencies = "250")
medicaid_gopb_orgs <- get_cobi_orgs(line_items = c("CBAA", "KPBA"))
orgs_2023 <- get_cobi_orgs(2023)
}
