% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi.R
\name{get_cobi_restricted_transfers}
\alias{get_cobi_restricted_transfers}
\title{Get COBI Restricted Fund and Account Transfers}
\usage{
get_cobi_restricted_transfers(
  years,
  agencies = NULL,
  line_items = NULL,
  financing_sources = NULL,
  is1x = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
tibble (modern data frame)
}
\description{
Get COBI Restricted Fund and Account Transfers
}
\examples{
restricted_fund_and_account_transfers <- get_cobi_restricted_transfers(c(2025, 2026))
gov_itf <- get_cobi_restricted_transfers(years = 2025, agencies = '060', financing_sources = '2480')
medicaid_aca_transfers <- get_cobi_restricted_transfers(years = 2025, line_items = '2252')
}
