% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi_data.R
\name{get_cobi_state_budget}
\alias{get_cobi_state_budget}
\title{Get COBI State Budget}
\usage{
get_cobi_state_budget(
  years,
  agencies = NULL,
  line_items = NULL,
  appr_units = NULL,
  financing_sources = NULL,
  is1x = NULL,
  bla = FALSE
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{appr_units}{character vector indicating appropriation unit(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}

\item{bla}{boolean value indicating whether to include Business-like Activities as part of the state budget}
}
\value{
tibble (modern data frame)
}
\description{
Get COBI State Budget
}
\examples{
ongoing_state_budget <- get_cobi_state_budget(c(2025, 2026), is1x = FALSE)
aco_federal <- get_cobi_state_budget(years = 2025, appr_units = 'KPBAC', financing_sources = '9750')
gopb_budget_with_bla <- get_cobi_state_budget(years = 2025, line_items = 'CBAA', bla = TRUE)
gov_from_gf <- get_cobi_state_budget(years = 2025, agencies = '060', financing_sources = '1000')
}
