% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi.R
\name{get_cobi_state_fund_expense}
\alias{get_cobi_state_fund_expense}
\title{Get COBI State Fund Expense}
\usage{
get_cobi_state_fund_expense(years, is1x = NULL)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
tibble (modern data frame)
}
\description{
Get COBI State Fund Expense
}
\examples{
state_fund_expense <- get_cobi_state_fund_expense(2025)
}
