% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cobi_data.R
\name{get_cobi_state_fund_withdrawal}
\alias{get_cobi_state_fund_withdrawal}
\title{Get COBI State Fund Withdrawal}
\usage{
get_cobi_state_fund_withdrawal(years, is1x = NULL)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
tibble (modern data frame)
}
\description{
This returns all appropriations made from unrestricted 'State Funds' including General Fund (1000), Income Tax Fund (2480), and Uniform School Fund (2400).
}
\examples{
state_fund_withdrawal <- get_cobi_state_fund_withdrawal(2025)
}
