% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi.R
\name{save_cobi_approps}
\alias{save_cobi_approps}
\title{Save COBI appropriations}
\usage{
save_cobi_approps(
  years,
  agencies = NULL,
  line_items = NULL,
  appr_units = NULL,
  financing_sources = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{appr_units}{character vector indicating appropriation unit(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI appropriations
}
\examples{
save_cobi_approps(2025)
save_cobi_approps(years = 2025, appr_units = 'KPBAC', financing_sources = '9750')
save_cobi_approps(years = c(2024, 2025), line_items = 'CBAA')
save_cobi_approps(years = 2025, line_items = '2105', financing_sources = '9800')
save_cobi_approps(years = c(2024, 2025), agencies = '060', financing_sources = '1000')
}
