% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_data.R
\name{save_cobi_cpf}
\alias{save_cobi_cpf}
\title{Save COBI Capital Project Funds}
\usage{
save_cobi_cpf(
  years,
  agencies = NULL,
  line_items = NULL,
  financing_sources = NULL,
  is1x = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI Capital Project Funds
}
\examples{
save_cobi_cpf(c(2025, 2026))
save_cobi_cpf(years = 2025, agencies = '102', financing_sources = '2480')
save_cobi_cpf(years = 2025, line_items = '2915')
}
