% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_ref.R
\name{save_cobi_expenditure_categories}
\alias{save_cobi_expenditure_categories}
\title{Save COBI expenditure categories}
\usage{
save_cobi_expenditure_categories()
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI expenditure categories
}
\examples{
save_cobi_expenditure_categories()
}
