% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_data.R
\name{save_cobi_fiduciary_funds}
\alias{save_cobi_fiduciary_funds}
\title{Save COBI Fiduciary Funds}
\usage{
save_cobi_fiduciary_funds(
  years,
  agencies = NULL,
  line_items = NULL,
  financing_sources = NULL,
  is1x = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI Fiduciary Funds
}
\examples{
save_cobi_fiduciary_funds(2025, line_items = '7208')
}
