% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_ref.R
\name{save_cobi_financing_sources}
\alias{save_cobi_financing_sources}
\title{Save COBI financing sources}
\usage{
save_cobi_financing_sources(year = NULL, restricted_only = NULL)
}
\arguments{
\item{year}{integer indicating fiscal year time stamp at which to snapshot active financing sources}

\item{restricted_only}{boolean value indicating whether to pull only restricted funds/accounts}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI financing sources
}
\examples{
save_cobi_financing_sources()
save_cobi_financing_sources(2024, TRUE)
}
