% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_data.R
\name{save_cobi_meta}
\alias{save_cobi_meta}
\title{Save COBI meta}
\usage{
save_cobi_meta(
  years,
  agencies = NULL,
  line_items = NULL,
  appr_units = NULL,
  categories = NULL,
  is1x = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{appr_units}{character vector indicating appropriation unit(s) of interest}

\item{categories}{character vector indicating financing source(s) or expenditure category(ies) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI meta
}
\examples{
save_cobi_meta(2025)
save_cobi_meta(years = 2025, appr_units = 'KPBAC', categories = '9750')
save_cobi_meta(years = c(2024, 2025), line_items = 'CBAA')
save_cobi_meta(years = 2025, line_items = '2105', categories = '9800')
save_cobi_meta(years = 2025, agencies = '060', categories = '1000')
save_cobi_meta(years = 2025, categories = 'AA')
}
