% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi.R
\name{save_cobi_orgs}
\alias{save_cobi_orgs}
\title{Save COBI orgs}
\usage{
save_cobi_orgs(year = NULL, agencies = NULL, line_items = NULL)
}
\arguments{
\item{year}{integer indicating year of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI orgs
}
\examples{
save_cobi_orgs()
save_cobi_orgs(agencies = "250")
save_cobi_orgs(line_items = c("CBAA", "KPBA"))
save_cobi_orgs(2023)
}
