% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_data.R
\name{save_cobi_proposed_expenditures}
\alias{save_cobi_proposed_expenditures}
\title{Get COBI proposed expenditures}
\usage{
save_cobi_proposed_expenditures(
  years,
  agencies = NULL,
  line_items = NULL,
  appr_units = NULL,
  expenditure_categories = NULL,
  is1x = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{appr_units}{character vector indicating appropriation unit(s) of interest}

\item{expenditure_categories}{character vector indicating expenditure category(ies) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Get COBI proposed expenditures
}
\examples{
save_cobi_proposed_expenditures(c(2025, 2026), agencies = '060')
save_cobi_proposed_expenditures(years = 2025, expenditure_categories = 'AA')
save_cobi_proposed_expenditures(years = 2025, line_items = 'KPBA', is1x = TRUE)
}
