% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_data.R
\name{save_cobi_restricted_transfers}
\alias{save_cobi_restricted_transfers}
\title{Save COBI Restricted Fund and Account Transfers}
\usage{
save_cobi_restricted_transfers(
  years,
  agencies = NULL,
  line_items = NULL,
  financing_sources = NULL,
  is1x = NULL
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI Restricted Fund and Account Transfers
}
\examples{
save_cobi_restricted_transfers(c(2025, 2026))
save_cobi_restricted_transfers(years = 2025, agencies = '060', financing_sources = '2480')
save_cobi_restricted_transfers(years = 2025, line_items = '2252')
}
