% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_data.R
\name{save_cobi_state_budget}
\alias{save_cobi_state_budget}
\title{Save COBI State Budget}
\usage{
save_cobi_state_budget(
  years,
  agencies = NULL,
  line_items = NULL,
  appr_units = NULL,
  financing_sources = NULL,
  is1x = NULL,
  bla = FALSE
)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{agencies}{character vector indicating agency(ies) of interest}

\item{line_items}{character vector indicating line item(s) of interest}

\item{appr_units}{character vector indicating appropriation unit(s) of interest}

\item{financing_sources}{character vector indicating financing source(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}

\item{bla}{boolean value indicating whether to include Business-like Activities as part of the state budget}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
Save COBI State Budget
}
\examples{
save_cobi_state_budget(c(2025, 2026), is1x = FALSE)
save_cobi_state_budget(years = 2025, appr_units = 'KPBAC', financing_sources = '9750')
save_cobi_state_budget(years = 2025, line_items = 'CBAA', bla = TRUE)
save_cobi_state_budget(years = 2025, agencies = '060', financing_sources = '1000')
}
