% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cobi_data.R
\name{save_cobi_state_funds_withdrawal}
\alias{save_cobi_state_funds_withdrawal}
\title{Save COBI State Funds Withdrawal}
\usage{
save_cobi_state_funds_withdrawal(years, is1x = NULL)
}
\arguments{
\item{years}{vector of integers indicating year(s) of interest}

\item{is1x}{boolean value indicating whether amount is one-time}
}
\value{
N/A; writes xlsx file to Downloads folder
}
\description{
This saves all appropriations made from unrestricted "State Funds" including General Fund (1000), Income Tax Fund (2480), and Uniform School Fund (2400) into an Excel file in the user's Downloads folder.
}
\examples{
save_cobi_state_funds_withdrawal(2025)
}
