% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopb-package.R
\docType{package}
\name{gopb-package}
\alias{gopb}
\alias{gopb-package}
\title{gopb: GOPB Toolkit}
\description{
Convenient functions to simplify and streamline common data acquisition, integration, manipulation, visualization, and publishing tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bolos-gopb/gopb}
  \item Report bugs at \url{https://github.com/bolos-gopb/gopb/issues}
}

}
\author{
\strong{Maintainer}: Jackson Bolos \email{jjbolos@utah.gov}

}
\keyword{internal}
